requirements.txt
joblib==1.3.2
matplotlib==3.8.1
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.1.0
Pillow==10.3.0
streamlit==1.31.0
tensorflow==1.12.0
xgboost==2.0.3
